Extract pendulum impl
